hv: move run_idle_thread to schedule file
move run_idle_thread and default_idle to schedule file as common
function.

Tracked-On: #8812
Acked-by: Wang, Yu1 <yu1.wang@intel.com>
