+++ hypervisor/common/hv_main.c
@@ -5,7 +5,6 @@
  */
 
 #include <asm/guest/vm.h>
-#include <asm/guest/vm_reset.h>
 #include <asm/guest/vmcs.h>
 #include <asm/guest/vmexit.h>
 #include <asm/guest/virq.h>
@@ -73,43 +72,4 @@ void vcpu_thread(struct thread_object *obj)
 
 		profiling_post_vmexit_handler(vcpu);
 	} while (1);
-}
-
-void default_idle(__unused struct thread_object *obj)
-{
-	uint16_t pcpu_id = get_pcpu_id();
-
-	while (1) {
-		if (need_reschedule(pcpu_id)) {
-			schedule();
-		} else if (need_offline(pcpu_id)) {
-			cpu_dead();
-		} else if (need_shutdown_vm(pcpu_id)) {
-			shutdown_vm_from_idle(pcpu_id);
-		} else {
-			cpu_do_idle();
-		}
-	}
-}
-
-void run_idle_thread(void)
-{
-	uint16_t pcpu_id = get_pcpu_id();
-	struct thread_object *idle = &per_cpu(idle, pcpu_id);
-	struct sched_params idle_params = {0};
-	char idle_name[16];
-
-	snprintf(idle_name, 16U, "idle%hu", pcpu_id);
-	(void)strncpy_s(idle->name, 16U, idle_name, 16U);
-	idle->pcpu_id = pcpu_id;
-	idle->thread_entry = default_idle;
-	idle->switch_out = NULL;
-	idle->switch_in = NULL;
-	idle_params.prio = PRIO_IDLE;
-	init_thread_data(idle, &idle_params);
-
-	run_thread(idle);
-
-	/* Control should not come here */
-	cpu_dead();
 }
