+++ hypervisor/include/common/schedule.h
@@ -148,6 +148,7 @@ void yield_current(void);
 void schedule(void);
 
 void arch_send_reschedule_request(uint16_t pcpu_id);
+void arch_default_idle(__unused struct thread_object *obj);
 void arch_switch_to(void *prev_sp, void *next_sp);
 void run_idle_thread(void);
 #endif /* SCHEDULE_H */
